# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code, e.g. 'not_found'.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context, e.g. field validation errors.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable explanation; safe to show to users.",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Correlates the response with server logs.",
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GoogleStartIn": {
        "additionalProperties": false,
        "properties": {
          "accept_terms": {
            "default": false,
            "description": "Required when creating an account.",
            "title": "Accept Terms",
            "type": "boolean"
          },
          "age_confirmed": {
            "default": false,
            "description": "As for /otp/verify: required for a new account or one without a recorded age confirmation (ADR 0009). Checked when Google sends the person back.",
            "title": "Age Confirmed",
            "type": "boolean"
          },
          "next": {
            "anyOf": [
              {
                "pattern": "^/([A-Za-z0-9\\-._~!$&'()*+,;=:@%?][A-Za-z0-9\\-._~!$&'()*+,;=:@%/?]{0,198})?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path on this site to return to after signing in (default /home).",
            "title": "Next"
          }
        },
        "title": "GoogleStartIn",
        "type": "object"
      },
      "GoogleStartOut": {
        "properties": {
          "authorization_url": {
            "description": "Send the browser here (Google's sign-in page).",
            "title": "Authorization Url",
            "type": "string"
          }
        },
        "required": [
          "authorization_url"
        ],
        "title": "GoogleStartOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OtpRequestIn": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "OtpRequestIn",
        "type": "object"
      },
      "OtpRequestOut": {
        "description": "Identical for every address, so it reveals nothing about which accounts exist.",
        "properties": {
          "expires_in_seconds": {
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "message": {
            "default": "If that address can receive email, a sign-in code is on its way.",
            "title": "Message",
            "type": "string"
          },
          "status": {
            "const": "sent",
            "default": "sent",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "expires_in_seconds"
        ],
        "title": "OtpRequestOut",
        "type": "object"
      },
      "OtpVerifyIn": {
        "additionalProperties": false,
        "properties": {
          "accept_terms": {
            "default": false,
            "description": "Required when creating an account: the user accepts the terms.",
            "title": "Accept Terms",
            "type": "boolean"
          },
          "age_confirmed": {
            "default": false,
            "description": "Self-declaration that the user is 18 or older (ADR 0009). Must be true to create an account, and to sign in to an account that has no recorded age confirmation.",
            "title": "Age Confirmed",
            "type": "boolean"
          },
          "code": {
            "description": "The 6-digit code from the email.",
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "email": {
            "maxLength": 254,
            "minLength": 3,
            "pattern": "^[^@\\s]+@[^@\\s]+\\.[^@\\s]+$",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email",
          "code"
        ],
        "title": "OtpVerifyIn",
        "type": "object"
      },
...
      "SessionList": {
        "properties": {
          "items": {
            "description": "Most recently used first.",
            "items": {
              "$ref": "#/components/schemas/SessionOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SessionList",
        "type": "object"
      },
      "SessionOut": {
        "description": "A signed-in device. Never its IP address or the full user agent.",
        "properties": {
          "created_at": {
            "description": "When this device signed in.",
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current": {
            "description": "True for the device making this request.",
            "title": "Current",
            "type": "boolean"
          },
          "device": {
            "description": "A short summary, e.g. \"Chrome on Windows\".",
            "title": "Device",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen_at": {
            "description": "Last used (updated at most hourly).",
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "device",
          "current",
          "created_at",
          "last_seen_at"
        ],
        "title": "SessionOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "email_verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email Verified At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_moderator": {
            "default": false,
            "description": "May use the moderation page and /api/v1/moderation.",
            "title": "Is Moderator",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "status": {
            "default": "active",
            "description": "`paused`: hidden from matching and new intros; chats and sign-in carry on.",
            "enum": [
              "active",
              "paused"
            ],
            "title": "Status",
            "type": "string"
          },
          "terms_accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Accepted At"
          },
          "terms_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terms Version"
          }
        },
        "required": [
          "id",
          "email",
          "created_at",
          "email_verified_at",
          "last_login_at",
          "terms_version",
          "terms_accepted_at"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Cynergi API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/google/callback": {
      "get": {
        "description": "Verifies the attempt and Google's ID token, then signs in (session cookie) and\nredirects to ``next``. Any failure redirects to ``/login?error=<code>`` with a stable\ncode (``google_state_invalid``, ``google_cancelled``, ``google_failed``,\n``email_not_allowed``, ``consent_required``, ...); details go to the audit log only.",
        "operationId": "google_callback_api_v1_auth_google_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "To `next`, or to /login?error=<code>"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Google sends the browser back here",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/google/start": {
      "post": {
        "description": "Returns Google's sign-in URL and sets a short-lived httpOnly cookie that ties the\nattempt to this browser. 404 ``google_signin_unavailable`` when Google is off.",
        "operationId": "google_start_api_v1_auth_google_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoogleStartOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start Google sign-in",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/logout-others": {
      "post": {
        "operationId": "logout_others_api_v1_auth_logout_others_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          }
        },
        "summary": "Sign out every other device",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/otp/request": {
      "post": {
        "description": "Send a 6-digit code (valid 10 minutes) to the address.\n\nThe response is the same whether or not an account exists. Requesting a new code\ninvalidates any earlier unused one. Rate-limited per address and per IP (429).",
        "operationId": "request_code_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Email a sign-in code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "description": "Verify the code and start a session (cookie). Creates the account on first use,\nwhich requires ``age_confirmed`` and ``accept_terms``; an account without a recorded age\nconfirmation needs ``age_confirmed`` too. Five wrong guesses lock a code.",
        "operationId": "verify_code_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Sign in with a code",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions": {
      "get": {
        "description": "Every device still signed in, most recently used first; ``current`` marks this one.",
        "operationId": "sessions_api_v1_auth_sessions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          }
        },
        "summary": "My signed-in devices",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/sessions/{session_id}": {
      "delete": {
        "description": "Signing out the current device this way also clears its cookies, like ``/logout``.",
        "operationId": "revoke_session_api_v1_auth_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not signed in."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not one of your devices, or already signed out (`session_not_found`)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sign out one device",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}